refactor(core): split numeric style scale sequential from identity
Extract map helpers and sequential/binned resolution from
scale-style-numeric.ts so identity resolution and the public
resolveNumericStyleScale orchestrator stay thin.
